+++ src/Modules/Tours/Explorer.Tours.Core/Domain/Problem.cs
@@ -13,10 +13,10 @@ namespace Explorer.Tours.Core.Domain
         public string Priority { get; private set; }
         public DateTime Date { get; private set; }
         public string Description { get; private set; }
-        public int TourId { get; private set; }
-        public int TouristId { get; private set; }
+        public long TourId { get; private set; }
+        public long TouristId { get; private set; }
 
-        public Problem(string category, string priority, DateTime date,string description, int tourId, int touristId)
+        public Problem(string category, string priority, DateTime date,string description, long tourId, long touristId)
         {
             this.Category = category;
             this.Priority = priority;
